# docs/Worth_It_OpenAPI_v1.0.yaml
openapi: 3.1.0
info:
  title: Worth It API
  version: 1.0.0
  description: Internal REST API contract for the Worth It mobile app MVP. All monetary amounts use integer minor
    units plus an ISO 4217 currency code.
servers:
- url: https://api.worthit.app/v1
  description: Production
- url: https://staging-api.worthit.app/v1
  description: Staging
security:
- bearerAuth: []
...
paths:
...
  /webhooks/revenuecat:
    post:
      summary: Receive RevenueCat entitlement events
      tags:
      - Webhooks
      responses:
        '204':
          description: Accepted
        '401':
          description: Invalid signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      security:
      - webhookSignature: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
...